pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vmware-monitor"
version = "0.1.0"
description = "VMware vCenter/ESXi read-only monitoring tool. No destructive operations — code-level enforced."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Wei Zhou / 周崴" }]
dependencies = [
    "pyvmomi>=8.0.3.0",
    "pyaml>=24.0",
    "typer>=0.12",
    "rich>=13.0",
    "apscheduler>=3.10,<4.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "mcp[cli]>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
vmware-monitor = "vmware_monitor.cli:app"
vmware-monitor-mcp = "mcp_server.server:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require vCenter)",
]
